Add fsuid/fsgid around the create operation
Make sure the user's uid/gid is accounted for the create operation
performed.

Now that we are setting the fsuid/fsgid to the real user's uid/gid, we do
not have to jump hoops to make sure the setgid bit in the parent directory
is honored. So, get rid of that code.

Signed-Off-by: Chandra Seetharaman <sekharan@us.ibm.com>
Signed-off-by: Venkateswararao Jujjuri (JV) <jvrao@linux.vnet.ibm.com>